Fixed Issues Caused by Refactoring Mistakes
Before commiting the last update, I compiling and played the game in Unreal Engine 4 to no issues, but it seems to have not actually compiled previously. As soon as I tried compiling and playing again, I noticed three issues; the Rocket's collision was always off, the Rocket could not move or rotate, and the asteroids were spawning on-screen. I fixed all of these for this commit.